+++ crates/env/src/call/call_builder.rs
@@ -65,7 +65,7 @@ where
     }
 }
 
-impl<E, Args, R> CallParams<E, Call<E>, Args, R>
+impl<E, Args, R> CallParams<E, CallV1<E>, Args, R>
 where
     E: Environment,
 {
@@ -88,6 +88,41 @@ where
     }
 }
 
+impl<E, Args, R> CallParams<E, Call<E>, Args, R>
+where
+    E: Environment,
+{
+    /// Returns the account ID of the called contract instance.
+    #[inline]
+    pub fn callee(&self) -> &E::AccountId {
+        &self.call_type.callee
+    }
+
+    /// Returns the chosen ref time limit for the called contract execution.
+    #[inline]
+    pub fn ref_time_limit(&self) -> u64 {
+        self.call_type.ref_time_limit
+    }
+
+    /// Returns the chosen proof time limit for the called contract execution.
+    #[inline]
+    pub fn proof_time_limit(&self) -> u64 {
+        self.call_type.proof_time_limit
+    }
+
+    /// Returns the chosen storage deposit limit for the called contract execution.
+    #[inline]
+    pub fn storage_deposit_limit(&self) -> Option<&E::Balance> {
+        self.call_type.storage_deposit_limit.as_ref()
+    }
+
+    /// Returns the transferred value for the called contract.
+    #[inline]
+    pub fn transferred_value(&self) -> &E::Balance {
+        &self.call_type.transferred_value
+    }
+}
+
 impl<E, Args, R> CallParams<E, DelegateCall<E>, Args, R>
 where
     E: Environment,
@@ -96,6 +131,45 @@ where
     #[inline]
     pub fn code_hash(&self) -> &E::Hash {
         &self.call_type.code_hash
+    }
+}
+
+impl<E, Args, R> CallParams<E, CallV1<E>, Args, R>
+where
+    E: Environment,
+    Args: scale::Encode,
+    R: scale::Decode,
+{
+    /// Invokes the contract with the given built-up call parameters.
+    ///
+    /// Returns the result of the contract execution.
+    ///
+    /// # Panics
+    ///
+    /// This method panics if it encounters an [`ink::env::Error`][`crate::Error`] or an
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`]. If you want to handle
+    /// those use the [`try_invoke`][`CallParams::try_invoke`] method instead.
+    pub fn invoke(&self) -> R {
+        crate::invoke_contract_v1(self)
+            .unwrap_or_else(|env_error| {
+                panic!("Cross-contract call failed with {env_error:?}")
+            })
+            .unwrap_or_else(|lang_error| {
+                panic!("Cross-contract call failed with {lang_error:?}")
+            })
+    }
+
+    /// Invokes the contract with the given built-up call parameters.
+    ///
+    /// Returns the result of the contract execution.
+    ///
+    /// # Note
+    ///
+    /// On failure this returns an outer [`ink::env::Error`][`crate::Error`] or inner
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`], both of which can be
+    /// handled by the caller.
+    pub fn try_invoke(&self) -> Result<ink_primitives::MessageResult<R>, crate::Error> {
+        crate::invoke_contract_v1(self)
     }
 }
 
@@ -204,11 +278,11 @@ where
 /// #     DefaultEnvironment,
 /// #     call::{build_call, Selector, ExecutionInput}
 /// # };
-/// # use ink_env::call::Call;
+/// # use ink_env::call::CallV1;
 /// # type AccountId = <DefaultEnvironment as Environment>::AccountId;
 /// # type Balance = <DefaultEnvironment as Environment>::Balance;
 /// build_call::<DefaultEnvironment>()
-///     .call(AccountId::from([0x42; 32]))
+///     .call_v1(AccountId::from([0x42; 32]))
 ///     .gas_limit(5000)
 ///     .transferred_value(10)
 ///     .exec_input(
@@ -236,11 +310,11 @@ where
 /// # use ::ink_env::{
 /// #     Environment,
 /// #     DefaultEnvironment,
-/// #     call::{build_call, Selector, ExecutionInput, Call},
+/// #     call::{build_call, Selector, ExecutionInput, CallV1},
 /// # };
 /// # type AccountId = <DefaultEnvironment as Environment>::AccountId;
 /// let my_return_value: i32 = build_call::<DefaultEnvironment>()
-///     .call_type(Call::new(AccountId::from([0x42; 32])))
+///     .call_type(CallV1::new(AccountId::from([0x42; 32])))
 ///     .gas_limit(5000)
 ///     .transferred_value(10)
 ///     .exec_input(
@@ -300,11 +374,11 @@ where
 /// #     DefaultEnvironment,
 /// #     call::{build_call, Selector, ExecutionInput}
 /// # };
-/// # use ink_env::call::Call;
+/// # use ink_env::call::CallV1;
 /// # type AccountId = <DefaultEnvironment as Environment>::AccountId;
 /// # type Balance = <DefaultEnvironment as Environment>::Balance;
 /// let call_result = build_call::<DefaultEnvironment>()
-///     .call(AccountId::from([0x42; 32]))
+///     .call_v1(AccountId::from([0x42; 32]))
 ///     .gas_limit(5000)
 ///     .transferred_value(10)
 ///     .try_invoke()
@@ -319,7 +393,7 @@ where
 #[allow(clippy::type_complexity)]
 pub fn build_call<E>() -> CallBuilder<
     E,
-    Unset<Call<E>>,
+    Unset<CallV1<E>>,
     Unset<ExecutionInput<EmptyArgumentList>>,
     Unset<ReturnType<()>>,
 >
@@ -335,17 +409,19 @@ where
     }
 }
 
-/// The default call type for cross-contract calls. Performs a cross-contract call to
+/// The legacy call type for cross-contract calls. Performs a cross-contract call to
 /// `callee` with gas limit `gas_limit`, transferring `transferred_value` of currency.
+///
+/// Calls into the original `call` host function.
 #[derive(Clone)]
-pub struct Call<E: Environment> {
+pub struct CallV1<E: Environment> {
     callee: E::AccountId,
     gas_limit: Gas,
     transferred_value: E::Balance,
 }
 
-impl<E: Environment> Call<E> {
-    /// Returns a clean builder for [`Call`].
+impl<E: Environment> CallV1<E> {
+    /// Returns a clean builder for [`CallV1`].
     pub fn new(callee: E::AccountId) -> Self {
         Self {
             callee,
@@ -355,13 +431,13 @@ impl<E: Environment> Call<E> {
     }
 }
 
-impl<E> Call<E>
+impl<E> CallV1<E>
 where
     E: Environment,
 {
     /// Sets the `gas_limit` for the current cross-contract call.
     pub fn gas_limit(self, gas_limit: Gas) -> Self {
-        Call {
+        CallV1 {
             callee: self.callee,
             gas_limit,
             transferred_value: self.transferred_value,
@@ -370,10 +446,35 @@ where
 
     /// Sets the `transferred_value` for the current cross-contract call.
     pub fn transferred_value(self, transferred_value: E::Balance) -> Self {
-        Call {
+        CallV1 {
             callee: self.callee,
             gas_limit: self.gas_limit,
             transferred_value,
+        }
+    }
+}
+
+/// The default call type for cross-contract calls, for calling into the latest `call_v2`
+/// host function. This adds the additional weight limit parameter `proof_time_limit` as
+/// well as `storage_deposit_limit`.
+#[derive(Clone)]
+pub struct Call<E: Environment> {
+    callee: E::AccountId,
+    ref_time_limit: u64,
+    proof_time_limit: u64,
+    storage_deposit_limit: Option<E::Balance>,
+    transferred_value: E::Balance,
+}
+
+impl<E: Environment> Call<E> {
+    /// Returns a clean builder for [`CallV1`].
+    pub fn new(callee: E::AccountId) -> Self {
+        Self {
+            callee,
+            ref_time_limit: Default::default(),
+            proof_time_limit: Default::default(),
+            storage_deposit_limit: None,
+            transferred_value: E::Balance::zero(),
         }
     }
 }
@@ -500,7 +601,23 @@ impl<E, CallType, Args, RetType> CallBuilder<E, Unset<CallType>, Args, RetType>
 where
     E: Environment,
 {
-    /// Prepares the `CallBuilder` for a cross-contract [`Call`].
+    /// Prepares the `CallBuilder` for a cross-contract [`CallV1`], calling into the
+    /// original `call` host function.
+    pub fn call_v1(
+        self,
+        callee: E::AccountId,
+    ) -> CallBuilder<E, Set<CallV1<E>>, Args, RetType> {
+        CallBuilder {
+            call_type: Set(CallV1::new(callee)),
+            call_flags: self.call_flags,
+            exec_input: self.exec_input,
+            return_type: self.return_type,
+            _phantom: Default::default(),
+        }
+    }
+
+    /// Prepares the `CallBuilder` for a cross-contract [`Call`] to the latest `call_v2`
+    /// host function.
     pub fn call(
         self,
         callee: E::AccountId,
@@ -529,7 +646,7 @@ where
     }
 }
 
-impl<E, Args, RetType> CallBuilder<E, Set<Call<E>>, Args, RetType>
+impl<E, Args, RetType> CallBuilder<E, Set<CallV1<E>>, Args, RetType>
 where
     E: Environment,
 {
@@ -537,7 +654,7 @@ where
     pub fn gas_limit(self, gas_limit: Gas) -> Self {
         let call_type = self.call_type.value();
         CallBuilder {
-            call_type: Set(Call {
+            call_type: Set(CallV1 {
                 callee: call_type.callee,
                 gas_limit,
                 transferred_value: call_type.transferred_value,
@@ -553,7 +670,7 @@ where
     pub fn transferred_value(self, transferred_value: E::Balance) -> Self {
         let call_type = self.call_type.value();
         CallBuilder {
-            call_type: Set(Call {
+            call_type: Set(CallV1 {
                 callee: call_type.callee,
                 gas_limit: call_type.gas_limit,
                 transferred_value,
@@ -562,6 +679,101 @@ where
             exec_input: self.exec_input,
             return_type: self.return_type,
             _phantom: Default::default(),
+        }
+    }
+}
+
+impl<E, Args, RetType> CallBuilder<E, Set<Call<E>>, Args, RetType>
+where
+    E: Environment,
+{
+    /// Switch to the original `call` host function API, which only allows the `gas_limit`
+    /// limit parameter (equivalent to the `ref_time_limit` in the latest `call_v2`).
+    ///
+    /// This method instance is used to allow usage of the generated call builder methods
+    /// for messages which initialize the builder with the original [`CallV1`] type.
+    pub fn call_v1(self) -> CallBuilder<E, Set<CallV1<E>>, Args, RetType> {
+        let call_type = self.call_type.value();
+        CallBuilder {
+            call_type: Set(CallV1 {
+                callee: call_type.callee,
+                gas_limit: call_type.ref_time_limit,
+                transferred_value: call_type.transferred_value,
+            }),
+            call_flags: self.call_flags,
+            exec_input: self.exec_input,
+            return_type: self.return_type,
+            _phantom: Default::default(),
+        }
+    }
+
+    /// Sets the `ref_time_limit` part of the weight limit for the current cross-contract
+    /// call.
+    ///
+    /// `ref_time` refers to the amount of computational time that can be
+    /// used for execution, in picoseconds. You can find more info
+    /// [here](https://use.ink/basics/cross-contract-calling/).
+    pub fn ref_time_limit(self, ref_time_limit: Gas) -> Self {
+        let call_type = self.call_type.value();
+        CallBuilder {
+            call_type: Set(Call {
+                ref_time_limit,
+                ..call_type
+            }),
+            ..self
+        }
+    }
+
+    /// Sets the `proof_time_limit` part of the weight limit for the current
+    /// cross-contract call.
+    ///
+    /// `proof_time` refers to the amount of storage in bytes that a transaction
+    /// is allowed to read. You can find more info
+    /// [here](https://use.ink/basics/cross-contract-calling/).
+    ///
+    /// **Note**
+    ///
+    /// This limit is only relevant for parachains, not for standalone chains which do not
+    /// require sending a Proof-of-validity to the relay chain.
+    pub fn proof_time_limit(self, proof_time_limit: Gas) -> Self {
+        let call_type = self.call_type.value();
+        CallBuilder {
+            call_type: Set(Call {
+                proof_time_limit,
+                ..call_type
+            }),
+            ..self
+        }
+    }
+
+    /// Sets the `storage_deposit_limit` for the current cross-contract call.
+    ///
+    /// The `storage_deposit_limit` specifies the amount of user funds that
+    /// can be charged for creating storage. You can find more info
+    /// [here](https://use.ink/basics/cross-contract-calling/).
+    pub fn storage_deposit_limit(self, storage_deposit_limit: E::Balance) -> Self {
+        let call_type = self.call_type.value();
+        CallBuilder {
+            call_type: Set(Call {
+                storage_deposit_limit: Some(storage_deposit_limit),
+                ..call_type
+            }),
+            ..self
+        }
+    }
+
+    /// Sets the `transferred_value` for the current cross-contract call.
+    ///
+    /// This value specifies the amount of user funds that are transferred
+    /// to the other contract with this call.
+    pub fn transferred_value(self, transferred_value: E::Balance) -> Self {
+        let call_type = self.call_type.value();
+        CallBuilder {
+            call_type: Set(Call {
+                transferred_value,
+                ..call_type
+            }),
+            ..self
         }
     }
 }
@@ -578,6 +790,23 @@ where
             exec_input: self.exec_input,
             return_type: self.return_type,
             _phantom: Default::default(),
+        }
+    }
+}
+
+impl<E, Args, RetType>
+    CallBuilder<E, Set<CallV1<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<RetType>>>
+where
+    E: Environment,
+{
+    /// Finalizes the call builder to call a function.
+    pub fn params(self) -> CallParams<E, CallV1<E>, Args, RetType> {
+        CallParams {
+            call_type: self.call_type.value(),
+            call_flags: self.call_flags,
+            _return_type: Default::default(),
+            exec_input: self.exec_input.value(),
+            _phantom: self._phantom,
         }
     }
 }
@@ -622,6 +851,28 @@ where
 }
 
 impl<E, RetType>
+    CallBuilder<
+        E,
+        Set<CallV1<E>>,
+        Unset<ExecutionInput<EmptyArgumentList>>,
+        Unset<RetType>,
+    >
+where
+    E: Environment,
+{
+    /// Finalizes the call builder to call a function.
+    pub fn params(self) -> CallParams<E, CallV1<E>, EmptyArgumentList, ()> {
+        CallParams {
+            call_type: self.call_type.value(),
+            call_flags: self.call_flags,
+            _return_type: Default::default(),
+            exec_input: Default::default(),
+            _phantom: self._phantom,
+        }
+    }
+}
+
+impl<E, RetType>
     CallBuilder<E, Set<Call<E>>, Unset<ExecutionInput<EmptyArgumentList>>, Unset<RetType>>
 where
     E: Environment,
@@ -657,6 +908,39 @@ where
             exec_input: Default::default(),
             _phantom: self._phantom,
         }
+    }
+}
+
+impl<E>
+    CallBuilder<
+        E,
+        Set<CallV1<E>>,
+        Unset<ExecutionInput<EmptyArgumentList>>,
+        Unset<ReturnType<()>>,
+    >
+where
+    E: Environment,
+{
+    /// Invokes the cross-chain function call.
+    ///
+    /// # Panics
+    ///
+    /// This method panics if it encounters an [`ink::env::Error`][`crate::Error`] or an
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`]. If you want to handle
+    /// those use the [`try_invoke`][`CallBuilder::try_invoke`] method instead.
+    pub fn invoke(self) {
+        self.params().invoke()
+    }
+
+    /// Invokes the cross-chain function call.
+    ///
+    /// # Note
+    ///
+    /// On failure this returns an outer [`ink::env::Error`][`crate::Error`] or inner
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`], both of which can be
+    /// handled by the caller.
+    pub fn try_invoke(self) -> Result<ink_primitives::MessageResult<()>, Error> {
+        self.params().try_invoke()
     }
 }
 
@@ -721,6 +1005,36 @@ where
     /// On failure this an [`ink::env::Error`][`crate::Error`] which can be handled by the
     /// caller.
     pub fn try_invoke(self) -> Result<ink_primitives::MessageResult<()>, Error> {
+        self.params().try_invoke()
+    }
+}
+
+impl<E, Args, R>
+    CallBuilder<E, Set<CallV1<E>>, Set<ExecutionInput<Args>>, Set<ReturnType<R>>>
+where
+    E: Environment,
+    Args: scale::Encode,
+    R: scale::Decode,
+{
+    /// Invokes the cross-chain function call and returns the result.
+    ///
+    /// # Panics
+    ///
+    /// This method panics if it encounters an [`ink::env::Error`][`crate::Error`] or an
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`]. If you want to handle
+    /// those use the [`try_invoke`][`CallBuilder::try_invoke`] method instead.
+    pub fn invoke(self) -> R {
+        self.params().invoke()
+    }
+
+    /// Invokes the cross-chain function call and returns the result.
+    ///
+    /// # Note
+    ///
+    /// On failure this returns an outer [`ink::env::Error`][`crate::Error`] or inner
+    /// [`ink::primitives::LangError`][`ink_primitives::LangError`], both of which can be
+    /// handled by the caller.
+    pub fn try_invoke(self) -> Result<ink_primitives::MessageResult<R>, Error> {
         self.params().try_invoke()
     }
 }